Expert_Agent/requirements.txt
# ============================================================
# Expert Agent - Minimal Production Requirements
# ============================================================

# --- Web Framework ---
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# --- Database ---
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# --- Caching ---
redis>=5.0.1

# --- Auth & Security ---
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# --- NLP & ML ---
spacy>=3.6.0
scikit-learn>=1.2.0
numpy>=1.24.0
# sentence-transformers pulls torch (~2GB). Install CPU-only torch first.
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0+cpu
sentence-transformers>=2.2.2

# --- Knowledge Graph ---
rdflib>=7.0.0

# --- Language Detection & Translation ---
langdetect>=1.0.9
deep-translator>=1.11.4

# --- Document Processing ---
pypdf>=3.0.0
python-docx>=1.0.0

# --- Utilities ---
requests>=2.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
watchdog>=3.0.0
